+++ packages/player/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - 2026-04-10
+
+### Changed
+
+- Use new `videoManifests` API for video playback instead of legacy playbackinfo endpoint
+- Shaka player updated to: 5.0.9
+
 ## [0.15.0] - 2026-03-03
 
 ### Changed